refactor(web): post-detail 컴포넌트 리팩토링
- post-detail 내에 있던 컴포넌트 개별 파일로 분리
