+++ apps/web/src/_pages/bulletin/index.ts
@@ -1,6 +1,9 @@
 export { BulletinInfoCard } from "./ui/bulletin-info-card";
 export { BulletinPopularPostCard } from "./ui/bulletin-popular-post-card";
 export { BulletinTable } from "./ui/bulletin-table";
+export { CommentItem } from "./ui/comment-item";
+export { CommentList } from "./ui/comment-list";
+export { PostArticle } from "./ui/post-article";
 export { PostCategoryGuideCard } from "./ui/post-category-guide-card";
-export { PostDetail } from "./ui/post-detail";
+export { PostInfoCard } from "./ui/post-info-card";
 export { PostWritingTipsCard } from "./ui/post-writing-tips-card";
